wip on create_plugin

+++ create_plugin.sh
@@ -1,7 +1,27 @@
-#!/bin/sh
+#!/bin/bash
+
+echo "Running unit tests"
+
+cd tests
+
+phpunit all.php
+# $? is a shell variable which stores the return code from what we just ran
+rc=$?
+
+# back to root dir
+cd ..
+
+if [[ $rc != 0 ]]; then
+	echo "phpunit failed - create plugin denied."
+	exit $rc
+fi
 
 # note: change to create a tag and push to git
 
+echo "Removing old plugin"
 rm thehubsa-plugin.zip
 
-git archive --format=zip HEAD -o thehubsa-plugin.zip 
+echo "Creating new plugin"
+git archive --format=zip HEAD -o thehubsa-plugin.zip
+
+echo "Done"
